back_end/translatorAPI/requirements.txt
pika
flask
flask-cors
python-keycloak